requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.7.4
python-telegram-bot==13.15
pytz==2024.1
six==1.16.0
tornado==6.1
tzdata==2024.1
tzlocal==5.2
